requirements.txt
pytest==7.4.3
selenium==4.15.2
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-ordering==0.6
PyYAML==6.0.1 